+++ translations/el_GR.ts
@@ -1532,6 +1532,10 @@ If it&apos;s partially checked, the global setting in Code Edit will be used.</s
         <source>WakaTime</source>
         <translation>WakaTime</translation>
     </message>
+    <message>
+        <source>Stopwatch</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>SettingsInfo</name>
@@ -2662,6 +2666,28 @@ It can be empty if you have the global wakatime config file ~/.wakatime.cfg.</so
         <comment>the anchor of Enable proxy for checking updates on the corresponding page of https://cpeditor.org/docs/preferences</comment>
         <translation>Η άγκυρα για την επιτροπή του μεταξούσιου για έλεγχο ενημερώσεις στην αντίστοίχη σελίδα https://cpeditor.org/docs/preferences</translation>
     </message>
+    <message>
+        <source>Display Stopwatch</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Start/Stop stopwatch on tab switch</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>When switching to a different tab, automatically start the stopwatch
+on the current tab and pause the stopwatch on the previous tab.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Show stopwatch result only when the button is pressed</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Hide the time of the stopwatch and only show the time when the &quot;Show&quot; button is pressed.
+This may reduce distractions caused by stopwatch updates.</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>ShortcutItem</name>
@@ -2873,6 +2899,33 @@ It can be empty if you have the global wakatime config file ~/.wakatime.cfg.</so
     <message>
         <source>Expected</source>
         <translation>Αναμενόμενα</translation>
+    </message>
+</context>
+<context>
+    <name>Widgets::Stopwatch</name>
+    <message>
+        <source>Start</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Pause</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Resume</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Reset</source>
+        <translation type="unfinished">Επαναφόρτωση</translation>
+    </message>
+    <message>
+        <source>Stopwatch</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Show</source>
+        <translation type="unfinished"></translation>
     </message>
 </context>
 <context>